+++ NavigationCorePackage/Tests/ExtasyNavigationCoreTests/WaypointDataTests.swift
@@ -1,0 +1,10 @@
+import XCTest
+@testable import ExtasyNavigationCore
+
+final class WaypointDataTests: XCTestCase {
+    func testResetIsCallableAndIdempotent() {
+        var data = WaypointData()
+        data.reset()
+        data.reset()
+    }
+}